Module gtfs_realtime::transit_realtime

Modules

alert

Nested message and enum types in Alert.

feed_header

Nested message and enum types in FeedHeader.

translated_string

Nested message and enum types in TranslatedString.

trip_descriptor

Nested message and enum types in TripDescriptor.

trip_update

Nested message and enum types in TripUpdate.

vehicle_position

Nested message and enum types in VehiclePosition.

Structs

Alert

An alert, indicating some sort of incident in the public transit network.

EntitySelector

A selector for an entity in a GTFS feed.

FeedEntity

A definition (or update) of an entity in the transit feed.

FeedHeader

Metadata about a feed, included in feed messages.

FeedMessage

The contents of a feed message. A feed is a continuous stream of feed messages. Each message in the stream is obtained as a response to an appropriate HTTP GET request. A realtime feed is always defined with relation to an existing GTFS feed. All the entity ids are resolved with respect to the GTFS feed. Note that “required” and “optional” as stated in this file refer to Protocol Buffer cardinality, not semantic cardinality. See reference.md at https://github.com/google/transit/tree/master/gtfs-realtime for field semantic cardinality.

Position

A position.

TimeRange

A time interval. The interval is considered active at time ‘t’ if ‘t’ is greater than or equal to the start time and less than the end time.

TranslatedString

An internationalized message containing per-language versions of a snippet of text or a URL. One of the strings from a message will be picked up. The resolution proceeds as follows:

TripDescriptor

A descriptor that identifies an instance of a GTFS trip, or all instances of a trip along a route.

TripUpdate

Realtime update of the progress of a vehicle along a trip. Depending on the value of ScheduleRelationship, a TripUpdate can specify:

VehicleDescriptor

Identification information for the vehicle performing the trip.

VehiclePosition

Realtime positioning information for a given vehicle.